# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/log/page": {
      "get": {
        "operationId": "LogController_getLogs",
        "summary": "Get all log items by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "log"
        ]
      }
    },
    "/log/type": {
      "get": {
        "operationId": "LogController_getLogsByType",
        "summary": "Get all log items by type",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "description": "日志类型",
            "schema": {
              "enum": [
                "Error",
                "Message"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "log"
        ]
      }
    },
...
    "/page": {
      "post": {
        "operationId": "PageController_create",
        "summary": "create one new page",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/page": {
      "get": {
        "operationId": "PageController_findAll",
        "summary": "find page list by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasAuth",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/{id}": {
      "get": {
        "operationId": "PageController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      },
      "patch": {
        "operationId": "PageController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/remove{id}": {
      "delete": {
        "operationId": "PageController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/enable/{id}": {
      "patch": {
        "operationId": "PageController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
...
    "/post/add": {
      "post": {
        "operationId": "PostController_create",
        "summary": "create one new post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/post/page": {
      "get": {
        "operationId": "PostController_findAll",
        "summary": "get all post by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "R",
                "O",
                "T"
              ],
              "type": "string"
            }
          },
          {
            "name": "visibleRange",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "P",
                "O",
                "L"
              ],
              "type": "string"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
...
    "/post/{id}": {
      "get": {
        "operationId": "PostController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Post"
        ]
      },
      "patch": {
        "operationId": "PostController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/post/remove/{id}": {
      "delete": {
        "operationId": "PostController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
    "/post/enable/{id}": {
      "patch": {
        "operationId": "PostController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Post"
        ]
      }
    },
...
    "/post-content/update": {
      "post": {
        "operationId": "PostContentController_update",
        "summary": "update post content",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "The id of the post content",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewPostContentCopy"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Post-Content"
        ]
      }
    },
    "/post-content/draft": {
      "post": {
        "operationId": "PostContentController_createDraft",
        "summary": "create one new draft",
        "parameters": [
          {
            "name": "postId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewPostContentCopy"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Post-Content"
        ]
      }
    }
  },
  "info": {
    "title": "My API",
    "description": "The description of my API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "scheme": "basic",
        "name": "access-token",
        "description": "user token authentication",
        "in": "headers"
      }
    },
    "schemas": {
...
      "CreatePageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "page title",
            "minLength": 2,
            "maxLength": 20
          },
          "path": {
            "type": "string",
            "description": "page path",
            "minLength": 4,
            "maxLength": 45
          },
          "sort": {
            "type": "number",
            "description": "page sort order",
            "minimum": 0,
            "maximum": 999
          },
          "component": {
            "type": "string",
            "description": "component path",
            "minLength": 4,
            "maxLength": 45
          },
          "isIframe": {
            "type": "string",
            "description": "is it iframe",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "type": {
            "type": "string",
            "description": "page type",
            "default": "P",
            "enum": [
              "P",
              "B"
            ]
          },
          "isLink": {
            "type": "string",
            "description": "is it external link",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "visible": {
            "type": "string",
            "description": "visible",
            "default": "0",
            "enum": [
              "0",
              "1"
            ]
          },
          "redirect": {
            "type": "string",
            "description": "redirect path",
            "minLength": 4,
            "maxLength": 45
          },
          "prefixIcon": {
            "type": "string",
            "description": "prefix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "suffixIcon": {
            "type": "string",
            "description": "suffix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "minLength": 4,
            "maxLength": 50
          },
          "auth": {
            "type": "string",
            "description": "auth id"
          }
        },
        "required": [
          "title",
          "path",
          "sort",
          "component",
          "isIframe",
          "type",
          "isLink",
          "visible",
          "redirect",
          "prefixIcon",
          "suffixIcon",
          "remark",
          "auth"
        ]
      },
      "UpdatePageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "page title",
            "minLength": 2,
            "maxLength": 20
          },
          "path": {
            "type": "string",
            "description": "page path",
            "minLength": 4,
            "maxLength": 45
          },
          "sort": {
            "type": "number",
            "description": "page sort order",
            "minimum": 0,
            "maximum": 999
          },
          "component": {
            "type": "string",
            "description": "component path",
            "minLength": 4,
            "maxLength": 45
          },
          "isIframe": {
            "type": "string",
            "description": "is it iframe",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "type": {
            "type": "string",
            "description": "page type",
            "default": "P",
            "enum": [
              "P",
              "B"
            ]
          },
          "isLink": {
            "type": "string",
            "description": "is it external link",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "visible": {
            "type": "string",
            "description": "visible",
            "default": "0",
            "enum": [
              "0",
              "1"
            ]
          },
          "redirect": {
            "type": "string",
            "description": "redirect path",
            "minLength": 4,
            "maxLength": 45
          },
          "prefixIcon": {
            "type": "string",
            "description": "prefix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "suffixIcon": {
            "type": "string",
            "description": "suffix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "minLength": 4,
            "maxLength": 50
          },
          "auth": {
            "type": "string",
            "description": "auth id"
          }
        },
        "required": [
          "title",
          "path",
          "sort",
          "component",
          "isIframe",
          "type",
          "isLink",
          "visible",
          "redirect",
          "prefixIcon",
          "suffixIcon",
          "remark",
          "auth"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "title",
            "minLength": 3,
            "maxLength": 99
          },
          "type": {
            "type": "string",
            "description": "type",
            "enum": [
              "R",
              "O",
              "T"
            ]
          },
          "coverImage": {
            "type": "string",
            "description": "cover image"
          },
          "summary": {
            "type": "string",
            "description": "summary",
            "maxLength": 180
          },
          "visibleRange": {
            "type": "string",
            "description": "visible range",
            "enum": [
              "P",
              "O",
              "L"
            ]
          },
          "content": {
            "type": "string",
            "description": "content"
          },
          "category": {
            "type": "array",
            "description": "catgory list"
          },
          "tags": {
            "type": "array",
            "description": "tag list"
          }
        },
        "required": [
          "title",
          "type",
          "coverImage",
          "summary",
          "visibleRange",
          "content",
          "category",
          "tags"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateNewPostContentCopy": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "post content"
          },
          "postId": {
            "type": "string",
            "description": "post id"
          }
        },
        "required": [
          "content"
        ]
      }
    }
  }
}